New functionality
Adds settings to mod.json allowing the player to control which functions the mod runs under.

By default, the mod runs much more frequently including on all (auto)saves, loading save games, beginning of the first round in-mission, and a handful of other places were the functions are called.

Mod no longer runs post-contract as there is a point between completion of selecting salvage and prior to loading the main UI. The player can re-enable this by setting ResolveCompleteContract to true in mod.json.